+++ home/project/code/Physics_Based_Particle_Filtering.py
@@ -12,7 +12,6 @@ from pickle import TRUE
 from re import T
 from ssl import ALERT_DESCRIPTION_ILLEGAL_PARAMETER
 from tkinter.tix import Tree
-
 import rospy
 import threading
 import rospkg
@@ -25,7 +24,6 @@ from geometry_msgs.msg import Point,PointStamped,PoseStamped,Quaternion,Transfor
 import tf
 import tf.transformations as transformations
 from visualization_msgs.msg import Marker
-
 #pybullet
 from pyquaternion import Quaternion
 import pybullet as p
@@ -44,14 +42,12 @@ import pandas as pd
 import multiprocessing
 #from sksurgerycore.algorithms.averagequaternions import average_quaternions
 from quaternion_averaging import weightedAverageQuaternions
-
-'''
-physicsClient = p.connect(p.GUI)#or p.DIRECT for non-graphical version
-p.setAdditionalSearchPath(pybullet_data.getDataPath()) #optionally
-p.setGravity(0,0,-9.81)
-p.resetDebugVisualizerCamera(cameraDistance=2,cameraYaw=0,cameraPitch=-40,cameraTargetPosition=[0.5,-0.9,0.5])
-planeId = p.loadURDF("plane.urdf")
-'''
+#class in other files
+from Franka_robot import Franka_robot
+from Ros_listener import Ros_listener
+from Particle import Particle
+from InitialRealworldModel import InitialRealworldModel
+from InitialSimulationModel import InitialSimulationModel
 
 boss_obse_err_sum_df = pd.DataFrame()
 boss_PFPE_err_sum_df = pd.DataFrame()
@@ -91,488 +87,12 @@ boss_estDO_ori_y_df = pd.DataFrame(columns=['step','time','ang_y','alg'],index=[
 boss_estDO_ori_z_df = pd.DataFrame(columns=['step','time','ang_z','alg'],index=[])
 boss_estDO_ori_w_df = pd.DataFrame(columns=['step','time','ang_w','alg'],index=[])
 
-
-
-
-
-
-
-
 boss_obs_pose_PFPM = []
 boss_est_pose_PFPM = []
 boss_update_flag_obse = 0
 boss_update_flag_PFPE = 0
 boss_update_flag_PFPM = 0
 
-'''
-#load and set franka robot
-plane_id = p_visualisation.loadURDF("plane.urdf")
-real_franka_robot_start_pos = [0.127, -0.04, 0.03]
-real_franka_robot_start_orientation = p_visualisation.getQuaternionFromEuler([0,0,0])
-real_robot_id = p_visualisation.loadURDF(os.path.expanduser("~/project/data/bullet3-master/examples/pybullet/gym/pybullet_data/franka_panda/panda.urdf"),
-                                         real_franka_robot_start_pos,
-                                         real_franka_robot_start_orientation,
-                                         useFixedBase=1)
-
-
-#load and set object
-cylinder_real_object_start_pos = [0.567, -0.3642, 0.057]
-cylinder_real_object_start_orientation = p_visualisation.getQuaternionFromEuler([0,0,0])
-real_object_id = p_visualisation.loadURDF(os.path.expanduser("~/project/object/cylinder_object_small.urdf"),
-                                          cylinder_real_object_start_pos,
-                                          cylinder_real_object_start_orientation)
-                                          
-'''
-
-#Class of franka robot move
-class Franka_robot():
-    def __init__(self,franka_robot_id):
-        self.franka_robot_id = franka_robot_id
-    def fanka_robot_move(self,targetPositionsJoints):
-        self.setJoinVisual(self.franka_robot_id,targetPositionsJoints)
-
-    def setJoinVisual(self,robot, position):
-        #position[7] = 0.039916139
-        #position[8] = 0.039916139
-        num_joints = 9
-        for joint_index in range(num_joints):
-            if joint_index == 7 or joint_index == 8:
-                p_visualisation.resetJointState(robot,
-                                                joint_index+2,
-                                                targetValue=position[joint_index])
-            else:
-                p_visualisation.resetJointState(robot,
-                                                joint_index,
-                                                targetValue=position[joint_index])
-
-    def setJointPosition(self,robot, position):
-        #position[7] = 0.039916139
-        #position[8] = 0.039916139
-        num_joints = 9
-        for joint_index in range(num_joints):
-            if joint_index == 7 or joint_index == 8:
-                p_visualisation.setJointMotorControl2(robot,
-                                                      joint_index+2,
-                                                      p_visualisation.POSITION_CONTROL,
-                                                      targetPosition=position[joint_index])
-            else:
-                p_visualisation.setJointMotorControl2(robot,
-                                                      joint_index,
-                                                      p_visualisation.POSITION_CONTROL,
-                                                      targetPosition=position[joint_index])
-
-#Class of franka robot listen to info from ROS
-class Ros_listener():
-    def __init__(self):
-        self.joint_subscriber = rospy.Subscriber('/joint_states', JointState, self.joint_values_callback,queue_size=1)
-        if optitrack_working_flag == True:
-            self.robot_pose = rospy.Subscriber('/mocap/rigid_bodies/pandaRobot/pose',PoseStamped, self.robot_pose_callback,queue_size=10)
-            if object_cracker_flag == True:
-                self.object_pose = rospy.Subscriber('/mocap/rigid_bodies/cheezit/pose',PoseStamped, self.object_pose_callback,queue_size=10)
-            if object_soup_flag == True:
-                self.object_pose = rospy.Subscriber('/mocap/rigid_bodies/zisongsoup/pose',PoseStamped, self.object_pose_callback,queue_size=10)
-            self.base_pose = rospy.Subscriber('/mocap/rigid_bodies/baseofcheezit/pose', PoseStamped, self.base_of_cheezit_callback,queue_size=10)
-        elif optitrack_working_flag == False:
-            self.fake_opti_pose = rospy.Subscriber('/Opti_pose',PoseStamped, self.fake_optipose_callback,queue_size=10)
-        self.current_joint_values = [-1.57,0.0,0.0,-2.8,1.7,1.57,1.1]
-        self.robot_pos = [ 0.139080286026,
-                          -0.581342339516,
-                           0.0238141193986]
-        #x,y,z,w
-        self.robot_ori = [ 0.707254290581,
-                           0.0115503482521,
-                          -0.0140119809657,
-                          -0.706726074219]
-
-        self.object_pos = [ 0.504023790359,
-                           -0.214561194181,
-                            0.0601389780641]
-        #x,y,z,w
-        self.object_ori = [-0.51964700222,
-                           -0.476704657078,
-                            0.490200251342,
-                            0.512272834778]
-        #self.object_ori = [0,0,0,1]
-        self.fake_opti_pos = [ 0.504023790359,
-                           -0.214561194181,
-                            0.0601389780641]
-        #x,y,z,w
-        self.fake_opti_ori = [-0.51964700222,
-                           -0.476704657078,
-                            0.490200251342,
-                            0.512272834778]
-        #self.object_ori = [0,0,0,1]
-        
-        rospy.spin
-    def joint_values_callback(self, msg):
-        self.current_joint_values = list(msg.position)
-    def robot_pose_callback(self, data):
-        #pos
-        x_pos = data.pose.position.x
-        y_pos = data.pose.position.y
-        z_pos = data.pose.position.z
-        self.robot_pos = [x_pos,y_pos,z_pos]
-        #ori
-        x_ori = data.pose.orientation.x
-        y_ori = data.pose.orientation.y
-        z_ori = data.pose.orientation.z
-        w_ori = data.pose.orientation.w
-        self.robot_ori = [x_ori,y_ori,z_ori,w_ori]
-    def object_pose_callback(self, data):
-        #pos
-        x_pos = data.pose.position.x
-        y_pos = data.pose.position.y
-        z_pos = data.pose.position.z
-        self.object_pos = [x_pos,y_pos,z_pos]
-        #ori
-        x_ori = data.pose.orientation.x
-        y_ori = data.pose.orientation.y
-        z_ori = data.pose.orientation.z
-        w_ori = data.pose.orientation.w
-        self.object_ori = [x_ori,y_ori,z_ori,w_ori]
-    def base_of_cheezit_callback(self,data):
-        # pos
-        x_pos = data.pose.position.x
-        y_pos = data.pose.position.y
-        z_pos = data.pose.position.z
-        self.base_pos = [x_pos, y_pos, z_pos]
-        # ori
-        x_ori = data.pose.orientation.x
-        y_ori = data.pose.orientation.y
-        z_ori = data.pose.orientation.z
-        w_ori = data.pose.orientation.w
-        self.base_ori = [x_ori, y_ori, z_ori, w_ori]
-    def fake_optipose_callback(self,data):
-        # pos
-        x_pos = data.pose.position.x
-        y_pos = data.pose.position.y
-        z_pos = data.pose.position.z
-        self.fake_opti_pos = [x_pos, y_pos, z_pos]
-        # ori
-        x_ori = data.pose.orientation.x
-        y_ori = data.pose.orientation.y
-        z_ori = data.pose.orientation.z
-        w_ori = data.pose.orientation.w
-        self.fake_opti_ori = [x_ori, y_ori, z_ori, w_ori]
-#Class of particle's structure
-class Particle(object):
-    def __init__(self,x=0.0,y=0.0,z=0.0,x_angle=0.0,y_angle=0.0,z_angle=0.0,w=1.0,index=0,linearVelocity=[0,0,0],angularVelocity=[0,0,0]):
-        self.x = x
-        self.y = y
-        self.z = z
-        self.x_angle = x_angle
-        self.y_angle = y_angle
-        self.z_angle = z_angle
-        self.w = w
-        self.index = index
-        self.linearVelocity = linearVelocity
-        self.angularVelocity = angularVelocity
-    def as_pose(self):
-        return True
-
-#Class of initialize the real world model
-class InitialRealworldModel():
-    def __init__(self,joint_pos=0):
-        self.flag = 0
-        self.joint_pos = joint_pos
-    def initial_robot(self,robot_pos,robot_orientation = [0,0,0,1]):
-        #robot_orientation = p_visualisation.getQuaternionFromEuler(robot_euler)
-        real_robot_id = p_visualisation.loadURDF(os.path.expanduser("~/project/data/bullet3-master/examples/pybullet/gym/pybullet_data/franka_panda/panda.urdf"),
-                                                 robot_pos,
-                                                 robot_orientation,
-                                                 useFixedBase=1)
-
-        self.set_real_robot_JointPosition(p_visualisation,real_robot_id,self.joint_pos)
-        for i in range(240):
-            p_visualisation.stepSimulation()
-            #time.sleep(1./240.)
-
-        return real_robot_id
-    def initial_target_object(self,object_pos,object_orientation = [0,0,0,1]):
-        #object_orientation = p_visualisation.getQuaternionFromEuler(object_euler)
-        if object_cracker_flag == True:
-            real_object_id = p_visualisation.loadURDF(os.path.expanduser("~/project/object/cube/cheezit_obj_small_hor.urdf"),
-                                                      object_pos,
-                                                      object_orientation)
-        if object_soup_flag == True:
-            real_object_id = p_visualisation.loadURDF(os.path.expanduser("~/project/object/soup/camsoup_obj_small_hor.urdf"),
-                                                      object_pos,
-                                                      object_orientation)
-        p_visualisation.changeDynamics(real_object_id,-1,mass=0.380,lateralFriction = 0.5)
-        return real_object_id
-    def set_real_robot_JointPosition(self,pybullet_simulation_env,robot, position):
-        # print("Preparing the joint pose of the panda robot!")
-        #position[7] = 0.039916139
-        #position[8] = 0.039916139
-        #num_joints = 7
-        num_joints = 9
-        for joint_index in range(num_joints):
-            if joint_index == 7 or joint_index == 8:
-                pybullet_simulation_env.setJointMotorControl2(robot,
-                                                              joint_index+2,
-                                                              pybullet_simulation_env.POSITION_CONTROL,
-                                                              targetPosition=position[joint_index])
-            else:
-                pybullet_simulation_env.setJointMotorControl2(robot,
-                                                              joint_index,
-                                                              pybullet_simulation_env.POSITION_CONTROL,
-                                                              targetPosition=position[joint_index])
-#Class of initialize the simulation model
-class InitialSimulationModel():
-    def __init__(self,particle_num,real_robot_start_pos,real_robot_start_ori,noise_obj_pos,noise_obj_ang,pw_T_object_ori_dope):
-        self.particle_num = particle_num
-        self.noise_obj_pos = noise_obj_pos
-        self.noise_obj_ang = noise_obj_ang
-        self.real_robot_start_pos = real_robot_start_pos
-        self.real_robot_start_ori = real_robot_start_ori
-        self.pw_T_object_ori_dope = pw_T_object_ori_dope
-        self.particle_cloud = []
-        self.pybullet_particle_env_collection = []
-        self.fake_robot_id_collection = []
-        self.particle_no_visual_id_collection = []
-        self.particle_with_visual_id_collection =[]
-        self.noise_object_pose = []
-
-        self.particle_cloud_PM = []
-        self.pybullet_particle_env_collection_PM = []
-        self.particle_no_visual_id_collection_PM = []
-        self.particle_with_visual_id_collection_PM =[]
-
-    def initial_particle(self):
-        noise_obj_x = copy.deepcopy(self.noise_obj_pos[0])
-        noise_obj_y = copy.deepcopy(self.noise_obj_pos[1])
-        noise_obj_z = copy.deepcopy(self.noise_obj_pos[2])
-        noise_obj_pos = [noise_obj_x,noise_obj_y,noise_obj_z]
-        noise_obj_x_ang = copy.deepcopy(self.noise_obj_ang[0])
-        noise_obj_y_ang = copy.deepcopy(self.noise_obj_ang[1])
-        noise_obj_z_ang = copy.deepcopy(self.noise_obj_ang[2])
-        noise_obj_ang = [noise_obj_x_ang,noise_obj_y_ang,noise_obj_z_ang]
-
-        self.noise_object_pose = [noise_obj_x,noise_obj_y,noise_obj_z,noise_obj_x_ang,noise_obj_y_ang,noise_obj_z_ang]
-
-        for i in range(self.particle_num):
-            x,y,z,x_angle,y_angle,z_angle,new_quat = self.generate_random_pose(self.noise_object_pose,self.pw_T_object_ori_dope)
-            w = 1/self.particle_num
-            particle = Particle(x,y,z,x_angle,y_angle,z_angle,w,index=i)
-            self.particle_cloud.append(particle)
-
-    def generate_random_pose(self,noise_object_pose, pw_T_object_ori_dope):
-        angle = copy.deepcopy([noise_object_pose[3],noise_object_pose[4],noise_object_pose[5]])
-        quat = copy.deepcopy(pw_T_object_ori_dope)#x,y,z,w
-        quat_QuatStyle = Quaternion(x=quat[0],y=quat[1],z=quat[2],w=quat[3])#w,x,y,z
-        x = self.add_noise_to_init_par(noise_object_pose[0],boss_sigma_obs_x)
-        y = self.add_noise_to_init_par(noise_object_pose[1],boss_sigma_obs_y)
-        z = self.add_noise_to_init_par(noise_object_pose[2],boss_sigma_obs_z)
-        random_dir = random.uniform(0, 2*math.pi)
-        z_axis = random.uniform(-1,1)
-        x_axis = math.cos(random_dir) * math.sqrt(1 - z_axis ** 2)
-        y_axis = math.sin(random_dir) * math.sqrt(1 - z_axis ** 2)
-        angle_noise = self.add_noise_to_init_par(0,boss_sigma_obs_ang_init)
-        w_quat = math.cos(angle_noise/2.0)
-        x_quat = math.sin(angle_noise/2.0) * x_axis
-        y_quat = math.sin(angle_noise/2.0) * y_axis
-        z_quat = math.sin(angle_noise/2.0) * z_axis
-        ###nois_quat(w,x,y,z); new_quat(w,x,y,z)
-        nois_quat = Quaternion(x=x_quat,y=y_quat,z=z_quat,w=w_quat)
-        new_quat = nois_quat * quat_QuatStyle
-        ###pb_quat(x,y,z,w)
-        pb_quat = [new_quat[1],new_quat[2],new_quat[3],new_quat[0]]
-        new_angle = p_visualisation.getEulerFromQuaternion(pb_quat)
-        x_angle = new_angle[0]
-        y_angle = new_angle[1]
-        z_angle = new_angle[2]
-        return x,y,z,x_angle,y_angle,z_angle,pb_quat
-    def compute_estimate_pos_of_object(self, particle_cloud):
-        x_set = 0
-        y_set = 0
-        z_set = 0
-        w_set = 0
-
-        quaternions = []
-        qws = []
-        for index,particle in enumerate(particle_cloud):
-            x_set = x_set + particle.x * particle.w
-            y_set = y_set + particle.y * particle.w
-            z_set = z_set + particle.z * particle.w
-            q = p_visualisation.getQuaternionFromEuler([particle.x_angle, particle.y_angle, particle.z_angle])
-            qws.append(particle.w)
-            quaternions.append([q[0], q[1], q[2], q[3]])
-            w_set = w_set + particle.w
-
-        # q = average_quaternions(np.array(quaternions))
-        q = weightedAverageQuaternions(np.array(quaternions), np.array(qws))
-        x_angle, y_angle, z_angle = p_visualisation.getEulerFromQuaternion([q[0], q[1], q[2], q[3]])
-
-        return x_set/w_set,y_set/w_set,z_set/w_set,x_angle,y_angle,z_angle
-
-
-    def display_particle(self):
-        for index, particle in enumerate(self.particle_cloud):
-            visualize_particle_pos = [particle.x, particle.y, particle.z]
-            visualize_particle_angle = [particle.x_angle, particle.y_angle, particle.z_angle]
-            visualize_particle_orientation = p_visualisation.getQuaternionFromEuler(visualize_particle_angle)
-            if object_cracker_flag == True:
-                visualize_particle_Id = p_visualisation.loadURDF(os.path.expanduser("~/project/object/cube/cheezit_par_with_visual_small_PE_hor.urdf"),
-                                                                visualize_particle_pos,
-                                                                visualize_particle_orientation)
-            if object_soup_flag == True:
-                visualize_particle_Id = p_visualisation.loadURDF(os.path.expanduser("~/project/object/soup/camsoup_par_with_visual_small_PE_hor.urdf"),
-                                                                visualize_particle_pos,
-                                                                visualize_particle_orientation)
-            self.particle_with_visual_id_collection.append(visualize_particle_Id)
-    def display_particle_PM(self):
-        for index, particle in enumerate(self.particle_cloud_PM):
-            visualize_particle_pos = [particle.x, particle.y, particle.z]
-            visualize_particle_angle = [particle.x_angle, particle.y_angle, particle.z_angle]
-            visualize_particle_orientation = p_visualisation.getQuaternionFromEuler(visualize_particle_angle)
-            if object_cracker_flag == True:
-                visualize_particle_Id = p_visualisation.loadURDF(os.path.expanduser("~/project/object/cube/cheezit_par_with_visual_small_PM_hor.urdf"),
-                                                                visualize_particle_pos,
-                                                                visualize_particle_orientation)
-            if object_soup_flag == True:
-                visualize_particle_Id = p_visualisation.loadURDF(os.path.expanduser("~/project/object/soup/camsoup_par_with_visual_small_PM_hor.urdf"),
-                                                                visualize_particle_pos,
-                                                                visualize_particle_orientation)
-            self.particle_with_visual_id_collection_PM.append(visualize_particle_Id)
-
-    def initial_and_set_simulation_env(self,joint_of_robot):
-        for index, particle in enumerate(self.particle_cloud):
-            pybullet_simulation_env = bc.BulletClient(connection_mode=p.DIRECT)#DIRECT,GUI_SERVER
-            self.pybullet_particle_env_collection.append(pybullet_simulation_env)
-            if update_style_flag == "time":
-                pybullet_simulation_env.setTimeStep(change_sim_time)
-            pybullet_simulation_env.resetDebugVisualizerCamera(cameraDistance=1, cameraYaw=180, cameraPitch=-85,
-                                                       cameraTargetPosition=[0.5, 0.3, 0.2])
-            pybullet_simulation_env.setAdditionalSearchPath(pybullet_data.getDataPath())
-            pybullet_simulation_env.setGravity(0,0,-9.81)
-            fake_plane_id = pybullet_simulation_env.loadURDF("plane.urdf")
-            if task_flag == "2":
-                if object_cracker_flag == True:
-                    sim_base_id = pybullet_simulation_env.loadURDF(os.path.expanduser("~/project/object/cube/base_of_cheezit.urdf"),
-                                                                pw_T_base_pos,
-                                                                pw_T_base_ori,
-                                                                useFixedBase=1)
-                if object_soup_flag == True:
-                    sim_base_id = pybullet_simulation_env.loadURDF(os.path.expanduser("~/project/object/cube/base_of_cheezit.urdf"),
-                                                                pw_T_base_pos,
-                                                                pw_T_base_ori,
-                                                                useFixedBase=1)
-            fake_robot_start_pos = self.real_robot_start_pos
-            fake_robot_start_orientation = self.real_robot_start_ori
-            fake_robot_id = pybullet_simulation_env.loadURDF(os.path.expanduser("~/project/data/bullet3-master/examples/pybullet/gym/pybullet_data/franka_panda/panda.urdf"),
-                                                             fake_robot_start_pos,
-                                                             fake_robot_start_orientation,
-                                                             useFixedBase=1)
-            self.fake_robot_id_collection.append(fake_robot_id)
-
-            #set joint of fake robot
-            self.set_sim_robot_JointPosition(pybullet_simulation_env,fake_robot_id,joint_of_robot)
-
-            particle_no_visual_start_pos = [particle.x, particle.y, particle.z]
-            particle_no_visual_start_angle = [particle.x_angle, particle.y_angle, particle.z_angle]
-            particle_no_visual_start_orientation = pybullet_simulation_env.getQuaternionFromEuler(particle_no_visual_start_angle)
-            if object_cracker_flag == True:
-                particle_no_visual_id = pybullet_simulation_env.loadURDF(os.path.expanduser("~/project/object/cube/cheezit_par_no_visual_small_hor.urdf"),
-                                                                        particle_no_visual_start_pos,
-                                                                        particle_no_visual_start_orientation)
-            if object_soup_flag == True:
-                particle_no_visual_id = pybullet_simulation_env.loadURDF(os.path.expanduser("~/project/object/soup/camsoup_par_no_visual_small_hor.urdf"),
-                                                                        particle_no_visual_start_pos,
-                                                                        particle_no_visual_start_orientation)
-            
-            while True:
-                pybullet_simulation_env.stepSimulation()
-                flag = 0
-                contacts = pybullet_simulation_env.getContactPoints(bodyA=fake_robot_id, bodyB=particle_no_visual_id)
-                # pmin,pmax = pybullet_simulation_env.getAABB(particle_no_visual_id)
-                # collide_ids = pybullet_simulation_env.getOverlappingObjects(pmin,pmax)
-                # length = len(collide_ids)
-                for contact in contacts:
-                    contact_dis = contact[8]
-                    if contact_dis < -0.001:
-                        #print("detected contact during initialization. BodyA: %d, BodyB: %d, LinkOfA: %d, LinkOfB: %d", contact[1], contact[2], contact[3], contact[4])
-                        Px,Py,Pz,Px_angle,Py_angle,Pz_angle,P_quat = self.generate_random_pose(self.noise_object_pose,self.pw_T_object_ori_dope)
-                        pybullet_simulation_env.resetBasePositionAndOrientation(particle_no_visual_id,
-                                                                                [Px,Py,Pz],
-                                                                                P_quat)
-                        flag = 1
-                        particle.x = Px
-                        particle.y = Py
-                        particle.z = Pz
-                        particle.x_angle = Px_angle
-                        particle.y_angle = Py_angle
-                        particle.z_angle = Pz_angle
-                        break
-                if flag == 0:
-                    break
-            #pybullet_simulation_env.changeDynamics(particle_no_visual_id,-1,mass=3,lateralFriction = 0.75)
-            self.particle_no_visual_id_collection.append(particle_no_visual_id)
-        obj_est_set = self.compute_estimate_pos_of_object(self.particle_cloud)
-        return obj_est_set[0],obj_est_set[1],obj_est_set[2],obj_est_set[3],obj_est_set[4],obj_est_set[5]
-    def initial_and_set_simulation_env_PM(self,joint_of_robot):
-        self.particle_cloud_PM = copy.deepcopy(self.particle_cloud)
-        for index, particle in enumerate(self.particle_cloud_PM):
-            pybullet_simulation_env = bc.BulletClient(connection_mode=p.DIRECT) # GUI_SERVER, DIRECT
-            self.pybullet_particle_env_collection_PM.append(pybullet_simulation_env)
-            pybullet_simulation_env.setAdditionalSearchPath(pybullet_data.getDataPath())
-            pybullet_simulation_env.setGravity(0,0,-9.81)
-            fake_plane_id = pybullet_simulation_env.loadURDF("plane.urdf")
-
-            particle_no_visual_start_pos = [particle.x, particle.y, particle.z]
-            particle_no_visual_start_angle = [particle.x_angle, particle.y_angle, particle.z_angle]
-            particle_no_visual_start_orientation = pybullet_simulation_env.getQuaternionFromEuler(particle_no_visual_start_angle)
-            if object_cracker_flag == True:
-                particle_no_visual_id = pybullet_simulation_env.loadURDF(os.path.expanduser("~/project/object/cube/cheezit_par_no_visual_small_hor.urdf"),
-                                                                        particle_no_visual_start_pos,
-                                                                        particle_no_visual_start_orientation)
-            if object_soup_flag == True:
-                particle_no_visual_id = pybullet_simulation_env.loadURDF(os.path.expanduser("~/project/object/soup/camsoup_par_no_visual_small_hor.urdf"),
-                                                                        particle_no_visual_start_pos,
-                                                                        particle_no_visual_start_orientation)
-            #pybullet_simulation_env.changeDynamics(particle_no_visual_id,-1,mass=3,lateralFriction = 0.7)
-            self.particle_no_visual_id_collection_PM.append(particle_no_visual_id)
-        obj_est_set_PM = self.compute_estimate_pos_of_object(self.particle_cloud_PM)
-        return obj_est_set_PM[0],obj_est_set_PM[1],obj_est_set_PM[2],obj_est_set_PM[3],obj_est_set_PM[4],obj_est_set_PM[5]
-
-
-    def set_sim_robot_JointPosition(self,pybullet_simulation_env,robot, position):
-        #position[7] = 0.039916139
-        #position[8] = 0.039916139
-        #num_joints = 7
-        # num_joints = 9
-        # for joint_index in range(num_joints):
-        #     if joint_index == 7 or joint_index == 8:
-        #         pybullet_simulation_env.setJointMotorControl2(robot,
-        #                                                       joint_index+2,
-        #                                                       pybullet_simulation_env.POSITION_CONTROL,
-        #                                                       targetPosition=position[joint_index])
-        #     else:
-        #         pybullet_simulation_env.setJointMotorControl2(robot,
-        #                                                       joint_index,
-        #                                                       pybullet_simulation_env.POSITION_CONTROL,
-        #                                                       targetPosition=position[joint_index])
-        # for i in range(240):
-        #     pybullet_simulation_env.stepSimulation()
-            #time.sleep(1./240.)
-        num_joints = 9
-        for joint_index in range(num_joints):
-            if joint_index == 7 or joint_index == 8:
-                pybullet_simulation_env.resetJointState(robot,
-                                                joint_index+2,
-                                                targetValue=position[joint_index])
-            else:
-                pybullet_simulation_env.resetJointState(robot,
-                                                joint_index,
-                                                targetValue=position[joint_index])
-                
-    def add_noise_to_init_par(self,current_pos,sigma_init):
-        mean = current_pos
-        sigma = sigma_init
-        new_pos_is_added_noise = self.take_easy_gaussian_value(mean, sigma)
-        return new_pos_is_added_noise
-    def take_easy_gaussian_value(self,mean,sigma):
-        normal = random.normalvariate(mean, sigma)
-        return normal
 
 #Class of Particle Filter move
 class PFMove():
@@ -1889,8 +1409,8 @@ if __name__ == '__main__':
     visualisation_flag = True
     visualisation_mean = False
     visualisation_particle_flag = True
-    object_cracker_flag = False
-    object_soup_flag = True
+    object_cracker_flag = True
+    object_soup_flag = False
     optitrack_working_flag = True
     publish_opti_pose_for_inter_flag = False
     write_opti_pose_flag = False
@@ -1976,7 +1496,7 @@ if __name__ == '__main__':
         p_visualisation.resetDebugVisualizerCamera(cameraDistance=1,cameraYaw=180,cameraPitch=-85,cameraTargetPosition=[0.3,0.1,0.2])
     plane_id = p_visualisation.loadURDF("plane.urdf")   
     #build an object of class "Ros_listener"
-    ros_listener = Ros_listener()
+    ros_listener = Ros_listener(optitrack_working_flag, object_cracker_flag, object_soup_flag)
     #get pose info from DOPE
     listener = tf.TransformListener()
     print("before while loop")
@@ -2043,6 +1563,8 @@ if __name__ == '__main__':
         
         
         #compute and load the pose of optitrack_base
+        pw_T_base_pos = [0,0,0]
+        pw_T_base_ori = [0,0,0,1]
         if task_flag == "2":
             robot_T_base = compute_transformation_matrix(init_robot_pos, init_robot_ori, base_of_cheezit_pos, base_of_cheezit_ori)
             # input('Press [ENTER] to compute the pose of object in the pybullet world')
@@ -2107,7 +1629,7 @@ if __name__ == '__main__':
 
     #input('Press [ENTER] to initial real world model')
     #build an object of class "InitialRealworldModel"
-    real_world_object = InitialRealworldModel(ros_listener.current_joint_values)
+    real_world_object = InitialRealworldModel(ros_listener.current_joint_values, object_cracker_flag, object_soup_flag, p_visualisation)
     #initialize the real robot in the pybullet
     real_robot_id = real_world_object.initial_robot(robot_pos = pybullet_robot_pos,robot_orientation = pybullet_robot_ori)
     # initialize the real object in the pybullet
@@ -2118,10 +1640,14 @@ if __name__ == '__main__':
         contact_particle_id = real_world_object.initial_target_object(object_pos = pw_T_object_pos_dope,object_orientation = pw_T_object_ori_dope)
         
     # build an object of class "Franka_robot"
-    franka_robot = Franka_robot(real_robot_id)
+    franka_robot = Franka_robot(real_robot_id, p_visualisation)
     # input('Press [ENTER] to initial simulation world model')
     # initialize sim world
-    initial_parameter = InitialSimulationModel(particle_num, pybullet_robot_pos, pybullet_robot_ori, dope_obj_pos_init, dope_obj_ang_init, dope_obj_ori_init)
+    # initial_parameter = InitialSimulationModel(particle_num, pybullet_robot_pos, pybullet_robot_ori, dope_obj_pos_init, dope_obj_ang_init, dope_obj_ori_init)
+    initial_parameter = InitialSimulationModel(particle_num, pybullet_robot_pos, pybullet_robot_ori, dope_obj_pos_init, dope_obj_ang_init, dope_obj_ori_init,
+                                               pw_T_base_pos, pw_T_base_ori,
+                                               boss_sigma_obs_x, boss_sigma_obs_y, boss_sigma_obs_z, boss_sigma_obs_ang_init, p_visualisation,
+                                               update_style_flag, change_sim_time, task_flag, object_cracker_flag, object_soup_flag)
     initial_parameter.initial_particle() #only position of particle
     
     estimated_object_set = initial_parameter.initial_and_set_simulation_env(ros_listener.current_joint_values)
@@ -2586,77 +2112,7 @@ if __name__ == '__main__':
         # print(t_end_while - t_begin)
         if Flag is False:
             break
-        # write the error file
-        '''
-        if rospy.is_shutdown():
-            print("++++++++++++++++++++++++++++++++++++++++++++++++")
-            if update_style_flag == "pose":
-                if task_flag == "1a":
-                    file_name_obse_pos = 'pose_scene1a_obse_err_pos.csv'
-                    file_name_PFPE_pos = 'pose_scene1a_PFPE_err_pos.csv'
-                    file_name_PFPM_pos = 'pose_scene1a_PFPM_err_pos.csv'
-                    file_name_obse_ang = 'pose_scene1a_obse_err_ang.csv'
-                    file_name_PFPE_ang = 'pose_scene1a_PFPE_err_ang.csv'
-                    file_name_PFPM_ang = 'pose_scene1a_PFPM_err_ang.csv'
-                elif task_flag == "1b":
-                    file_name_obse_pos = 'pose_scene1b_obse_err_pos.csv'
-                    file_name_PFPE_pos = 'pose_scene1b_PFPE_err_pos.csv'
-                    file_name_PFPM_pos = 'pose_scene1b_PFPM_err_pos.csv'
-                    file_name_obse_ang = 'pose_scene1b_obse_err_ang.csv'
-                    file_name_PFPE_ang = 'pose_scene1b_PFPE_err_ang.csv'
-                    file_name_PFPM_ang = 'pose_scene1b_PFPM_err_ang.csv'
-                elif task_flag == "2":
-                    file_name_obse_pos = 'pose_scene2_obse_err_pos.csv'
-                    file_name_PFPE_pos = 'pose_scene2_PFPE_err_pos.csv'
-                    file_name_PFPM_pos = 'pose_scene2_PFPM_err_pos.csv'
-                    file_name_obse_ang = 'pose_scene2_obse_err_ang.csv'
-                    file_name_PFPE_ang = 'pose_scene2_PFPE_err_ang.csv'
-                    file_name_PFPM_ang = 'pose_scene2_PFPM_err_ang.csv'
-            elif update_style_flag == "time":
-                if task_flag == "1a":
-                    file_name_obse_pos = 'time_scene1a_obse_err_pos.csv'
-                    file_name_PFPE_pos = 'time_scene1a_PFPE_err_pos.csv'
-                    file_name_PFPM_pos = 'time_scene1a_PFPM_err_pos.csv'
-                    file_name_obse_ang = 'time_scene1a_obse_err_ang.csv'
-                    file_name_PFPE_ang = 'time_scene1a_PFPE_err_ang.csv'
-                    file_name_PFPM_ang = 'time_scene1a_PFPM_err_ang.csv'
-                elif task_flag == "1b":
-                    file_name_obse_pos = 'time_scene1b_obse_err_pos.csv'
-                    file_name_PFPE_pos = 'time_scene1b_PFPE_err_pos.csv'
-                    file_name_PFPM_pos = 'time_scene1b_PFPM_err_pos.csv'
-                    file_name_obse_ang = 'time_scene1b_obse_err_ang.csv'
-                    file_name_PFPE_ang = 'time_scene1b_PFPE_err_ang.csv'
-                    file_name_PFPM_ang = 'time_scene1b_PFPM_err_ang.csv'
-                elif task_flag == "2":
-                    file_name_obse_pos = 'time_scene2_obse_err_pos.csv'
-                    file_name_PFPE_pos = 'time_scene2_PFPE_err_pos.csv'
-                    file_name_PFPM_pos = 'time_scene2_PFPM_err_pos.csv'
-                    file_name_obse_ang = 'time_scene2_obse_err_ang.csv'
-                    file_name_PFPE_ang = 'time_scene2_PFPE_err_ang.csv'
-                    file_name_PFPM_ang = 'time_scene2_PFPM_err_ang.csv'
-                elif task_flag == "3":
-                    file_name_obse_pos = 'time_scene3_obse_err_pos.csv'
-                    file_name_PFPE_pos = 'time_scene3_PFPE_err_pos.csv'
-                    file_name_PFPM_pos = 'time_scene3_PFPM_err_pos.csv'
-                    file_name_obse_ang = 'time_scene3_obse_err_ang.csv'
-                    file_name_PFPE_ang = 'time_scene3_PFPE_err_ang.csv'
-                    file_name_PFPM_ang = 'time_scene3_PFPM_err_ang.csv'
-            # boss_err_pos_df.to_csv('error_file/'+str(file_time)+file_name_pos,index=0,header=0,mode='a')
-            # boss_err_ang_df.to_csv('error_file/'+str(file_time)+file_name_ang,index=0,header=0,mode='a')
-            if run_PFPE_flag == True:
-                boss_obse_err_pos_df.to_csv('error_file/'+str(file_time)+file_name_obse_pos,index=0,header=0,mode='a')
-                boss_obse_err_ang_df.to_csv('error_file/'+str(file_time)+file_name_obse_ang,index=0,header=0,mode='a')
-                print("write obser file")
-                boss_PFPE_err_pos_df.to_csv('error_file/'+str(file_time)+file_name_PFPE_pos,index=0,header=0,mode='a')
-                boss_PFPE_err_ang_df.to_csv('error_file/'+str(file_time)+file_name_PFPE_ang,index=0,header=0,mode='a')
-                print("write PFPE file")
-                print("PE: Update frequency is: " + str(flag_update_num_PE))
-            if run_PFPM_flag == True:
-                boss_PFPM_err_pos_df.to_csv('error_file/'+str(file_time)+file_name_PFPM_pos,index=0,header=0,mode='a')
-                boss_PFPM_err_ang_df.to_csv('error_file/'+str(file_time)+file_name_PFPM_ang,index=0,header=0,mode='a')
-                print("write PFPM file")
-                print("PM: Update frequency is: " + str(flag_update_num_PM))
-        '''
+    
     p_visualisation.disconnect()
